Update finalise tests
