+++ tests/data/finalise_data.py
@@ -1,4 +1,2 @@
-cmd_line_before = "dwc_otg.lpm_enable=0 console=tty1 root=PARTUUID=f3934ed9-06 rootfstype=ext4 elevator=deadline fsck.repair=yes rootwait modules-load=dwc2,g_ether quiet splash plymouth.ignore-serial-consoles logo.nologo vt.global_cursor_default=0 loglevel=0 fbcon=map:2\n"  # noqa: E501
-cmd_line_after = "dwc_otg.lpm_enable=0 console=tty1 root=PARTUUID=f3934ed9-06 rootfstype=ext4 elevator=deadline fsck.repair=yes rootwait modules-load=dwc2,g_ether quiet splash plymouth.ignore-serial-consoles logo.nologo vt.global_cursor_default=0 loglevel=0"  # noqa: E501
 available_space_out = b"Avail\n1000000000000"
 available_space = "1000000000000"
